requirements.txt
numpy
tensorflow<2.0
Keras
Pillow
scikit-image
nltk
matplotlib
